[FAB-1335] Verify the setting of environ vars
This allows for the testing of the setting of environment
variables. It also adds some tests that can be used as
smoke tests for executing the functionality of the behave
tests to make sure things are working as expected.

This can be used by CI when code is pushed to this repository
for verifying changes without executing the full suite of tests.

Change-Id: Iba844490f8a9ca8503dae35d056b1ca467619df7
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>