[FAB-4684] Gossip-leader test with leader select
This tests the gossip-leader's ability to catch up in the case it was down
temporarily and there were blocks generated during that time. Leader is
set via gossip leader and leader-election is off.

Also takes care of [FAB-4683] and [FAB-4685], similar tests with
different methods of taking down the leader.

Change-Id: Id34234d2e8fb9fcc04e26320c0b2921fef1be8cc
Signed-off-by: Adnan Choudhury <adnan.choudhury@itpeoplecorp.com>