[FAB-9719] inconsistent txcounts from orderers
This is not a final solution. In this CR:
- enabled debugflag1 to get some more timely info
from OTE
- cleaned up some logging in OTE
- added debug options for runote.sh, and gather
memory data from host at completion of each test
- fixed couple problems that were causing the
tests to fail when invoked manually:
one was in the NL tool, to change the pathname
of the fabric binaries from build to .build, and
another was to change the OTE code when starting
Consumers, to adapt to newer grpc version.
- also updated submodules

Change-Id: I5d94fbeab0fe7f548cbe45f357aec9da52ccfd26
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>