[FAB-10989] Fix node calls for NodeSDK tests
The tests that use NodeSDK were failing due to
a code change that assumed client info was set
at an earlier time in the code. The test was
also having some problems running in vagrant.
This CR addresses both.

There is an issue with using npm in a vagrant
VM, where the node_modules directory needs to
be native on the machine and cannot be a synced
directory. With the change to sync the
"hyperledger" directory when using vagrant,
going up 3 levels gets us to a native machine
directory.

Change-Id: I1c7d9c5c93bdf8dc69b680d2d6c17f6abad68658
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>