FAB-12254 Add latest tag to thirdparty images.
There are no `latest` tagged thirdparty images
and the Behave tests are looking for the `latest`
image which doesn't exist as the thirdparty images
in the CI have a different tag. This change is to
pull the thirdparty images and tag to `latest` to
allow the normal testing in the Behave tests.

Change-Id: Ic838e0c30c72c7f065088391c7ca6d86046d801f
Signed-off-by: Sambhav Nidamarty <sambhavdutt@gmail.com>