[FAB-12629] runPTE.sh on Mac Darwin OS
Fix PTE scripts to be portable. Pay special
attention to using "sed" command with "-i" option.
Also, the mac OS can not handle the "%N" nanosecs
formatting option when using "date" command, so
we avoid that when on a mac; it is nice detail to
have, but not critical to anything today anyways.

Change-Id: I65b8d20f6d4f3adbc7f6b0ef0947935a8b744341
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>