[FAB-12241] Running npm install at once in tests
Here "npm install" is performed at the
beginning of every test suite.

For example, After this change: The
smoke took 4m31.668s instead of
6m6.022s. Hope this will decrease time
for daily test also.

Care was taken to ensure that Node
modules are delete at the end of the
test suite

Removed space based on the comments.
Change-Id: I7759812743b4134fa9f9ff59d3c2092340d7ac67
Signed-off-by: ChandraLekhaChavva <chandra.chavva@itpeoplecorp.com>