[FAB-13596] PTE supports orderer load balance
Currently, PTE uses the orderer defined in
ordererID in the Service Credentials File for all
transactions. This is less flexibility for load
balance in sending transactions to all orderers.
This CR provides a additional method, round robin,
to achieve better load balance in orderer
transactions.

Current method:
User defines the orderer to be used for all
transaction in ordererID in the PTE Service
Credentials file for each org.  The defined
orderer must be included in the orderer list
in the same file.

Proposed method:
PTE will supply two methods of orderer selection:
1. UserDefined: same as current method, i.e. user
defines the orderer to be used for all transactions
for each org in the Service Credentials file
2. RoundRobin: PTE will assign an orderer from the
orderer list defined in the Service Credentials File
in a round robin fashion to each thread.  Note that
the orderer list should only include the orderers to
participate in the transactions in the list.

Change-Id: I827eaf7b987c9cc7523af03cc016fe702ec1bd1f
Signed-off-by: Dongming <dongming@ibm.com>