Remove Long Running Tests
Both the production and chaos suites need to be completely
refactored. In particular, the production suite needs to be
migrated to chaincodes that excersize more realistic production
functionality of fabric such as complex queries and private data.

The chaos suite needs to be reimplemented to do actual process
cancellation, i.e., `kubectl exec -it <pod name> -- kill -9 0`
this will do a forceful cancellation rather than allowing for
a graceful shutdown as even a forceful `kubectl delete -f` only
removes the pod from the API server registration and does not
kill the pod.

We will revisit the implemention in a later Jira.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>