Nominate Dave Enyeart as Maintainer and Retire Inactive Maintainers
Dave Enyeart is the core Hyperledger Fabric release manager. For the
last several years Dave has managed the priorities of not only Fabric
but Fabric-Test. Many if not most of the requirements for Fabric test
have been driven by requirements from Dave, or through conversations
with Dave. As such, Dave's nomination for maintainership of Fabric-Test
seems a no-brainer.

This change also removes the Release Mangers sections as Fabric-Test
does not manage releases so there is no need for this section to echo
what the core Fabric repo does.

Signed-off-by: Brett Logan <lindluni@github.com>