Add Daily Test Suite to GitHub Actions
Add a GitHub Workflow to run the daily test suite.
As part of this change, the upgrade scenarios were
split out into their own workflow, as they are
dependent on a different directory setup due to the
way the tests are run, i.e., checking out 1.4 branch
which has different functionality for handling the
chaincode path. To keep things clean, and not change
settings on the other tests that don't really affect
them, the upgrade tests were split into their own
workflow file.

Signed-off-by: Brett Logan <lindluni@github.com>