Add Daily Regression Suite to GitHub Actions
Add the Daily Regression Suite to GitHub Actions.
As part of this change the upgrade scenarios were
split out into their own workflow files as the setup
for them is different as they are driven by bash,
rather than being driven by Ginkgo. In an effort to
not do custom implementations on the non-upgrade tests
to satisfy the requirements of the upgrade tests,
they were split out into their own workflow.

Signed-off-by: Brett Logan <lindluni@github.com>