Remove indirect upgrade test
The indirect upgrade test (v1.4 to v2.0 to v2.2) does not work with Go 1.16
since it steps through release-2.0 which does not
work with Go 1.16.
This commit stops running the test on the environment that
was recently updated to Go 1.16.

There is still a direct upgrade test from v1.4 to v2.2 which works.
This is the recommended path for users (LTS release to LTS release).

Signed-off-by: David Enyeart <enyeart@us.ibm.com>